# docs/api-v1.openapi.yaml
openapi: 3.0.3
info:
  title: Arka01 — API móvil (v1)
  description: >
    Contrato de /api/v1, consumido por la app móvil Capacitor
    (ROADMAP_APLICACION_MOVIL_CAPACITOR.md). Autenticación por token Bearer
    de Laravel Sanctum (ver /auth/login, /auth/register). Cada endpoint reusa
    exactamente la misma lógica de negocio que la web equivalente, extraída
    a App\Services\* — nunca hay una regla de negocio distinta entre los dos
    canales, solo el formato de respuesta cambia (JSON acá, Inertia allá).
    Esta especificación cubre el contrato completo construido hasta la
    pasada "full backend" del 2026-08-28 (ver ROADMAP_APLICACION_MOVIL_PROGRESO.md)
    — no incluye pantallas que todavía no existen ni el catálogo de
    planes/pago manual, deliberadamente fuera de alcance.
  version: "1.0"
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /driver/profile:
    get:
      tags: [Conductor]
      summary: Perfil completo del conductor (para prellenar la pantalla de edición)
      responses:
        "200": { description: "{ profile } — profile es null si todavía no tiene uno" }
    post:
      tags: [Conductor]
      summary: Crear/editar el perfil de conductor (vehículo, tarifas, cobertura, documentos)
      description: >
        multipart/form-data si incluye archivos (profile_photo,
        identity_document, license_photo, police_record). Los campos de
        identidad del vehículo (marca, modelo, color, tipo, placa, año,
        capacidad, cajuela) quedan fijados después del primer guardado —
        el servidor ignora cualquier cambio a esos campos una vez existen.
        Cambiar datos "revisados" o subir documentos nuevos reabre
        verification_status a pending y desconecta al conductor.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [vehicle_make, vehicle_model, vehicle_color, vehicle_type, vehicle_plate, vehicle_year, passenger_capacity, has_trunk, rate_per_km]
              properties:
                country_code: { type: string }
                phone_local: { type: string }
                driver_type: { type: string, enum: [independent, public_transport] }
                vehicle_make: { type: string }
                vehicle_model: { type: string }
                vehicle_color: { type: string }
                vehicle_type: { type: string }
                vehicle_plate: { type: string }
                vehicle_year: { type: integer }
                passenger_capacity: { type: integer, minimum: 1, maximum: 8 }
                has_trunk: { type: boolean }
                vehicle_amenities: { type: array, items: { type: string } }
                rate_per_km: { type: number }
                minimum_fare: { type: number, nullable: true }
                max_request_distance_km: { type: integer, nullable: true }
                accepts_cash: { type: boolean }
                accepts_transfer: { type: boolean }
                has_insurance: { type: boolean }
                is_public: { type: boolean }
                profile_public: { type: boolean }
                profile_photo: { type: string, format: binary }
                identity_document: { type: string, format: binary }
                license_photo: { type: string, format: binary }
                police_record: { type: string, format: binary }
      responses:
        "200": { description: "{ profile }" }
        "422": { description: "Campo de vehículo faltante / documento obligatorio faltante para verificación / tarifa mínima por encima del tope de plataforma" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "Sanctum personal access token"